New component 'LectureTable', with Datatable functionality
